moved EthernetClientStream implementation to header
fixes compilation issues for Intel Galileo and Edison
as well as other boards that may not be fully compatible
with the Arduino Ethernet library
